+++ ZestyBiteWebAppSolution/Models/Entities/Payment.cs
@@ -7,7 +7,7 @@ public partial class Payment
 {
     public int PaymentId { get; set; }
 
-    public ulong PaymentMethod { get; set; }
+    public int PaymentMethod { get; set; }
 
     public virtual ICollection<Bill> Bills { get; set; } = new List<Bill>();
 }
