+++ ZestyBiteWebAppSolution/Models/Entities/Table.cs
@@ -9,13 +9,13 @@ public partial class Table
 
     public int TableCapacity { get; set; }
 
-    public ulong TableMaintenance { get; set; }
+    public int TableMaintenance { get; set; }
 
     public int ReservationId { get; set; }
 
     public int ItemId { get; set; }
 
-    public ulong TableType { get; set; }
+    public int TableType { get; set; }
 
     public string TableStatus { get; set; } = null!;
